mnt/outputs/github-portfolio/data-lakehouse-etl/requirements.txt
pyspark==3.5.2
delta-spark==3.2.0
kafka-python==2.0.2
great-expectations==0.18.15
boto3==1.34.144
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==17.0.0
